split out initPerf into own file
